Move Pulse, DistanceResolution, TimeResolution, LtotalRange into wavelength_lut_workflow_specs
These models are specific to the wavelength LUT workflow and have no
other callers. parameter_models.py is for generic, reusable building
blocks shared across multiple specs.

Also avoids the naming confusion with the identically-named scipp
workflow type keys in ess.reduce.unwrap.lut.
