+++ src/ess/livedata/parameter_models.py
@@ -239,51 +239,3 @@ def make_edges(*, model: EdgesModel, dim: str, unit: str) -> sc.Variable:
     return op(
         dim=dim, start=model.start, stop=model.stop, num=model.num_bins + 1, unit=unit
     )
-
-
-class Pulse(BaseModel):
-    """Source pulse properties.
-
-    The unit of ``frequency`` is hardcoded to Hz: there is no realistic case
-    where a neutron-source pulse frequency would be expressed in kHz/MHz, so
-    we skip the dropdown and keep the model tight.
-    """
-
-    frequency: float = Field(default=14.0, gt=0.0, description="Pulse frequency in Hz.")
-    stride: int = Field(
-        default=1,
-        ge=1,
-        description="Pulse stride (1 unless pulse-skipping is used).",
-    )
-
-    def get_period(self) -> sc.Variable:
-        """Return the pulse period as a scipp scalar."""
-        return sc.scalar(1.0 / self.frequency, unit='s')
-
-
-class DistanceResolution(BaseModel):
-    """Resolution of the distance axis in a TOF lookup table."""
-
-    value: float = Field(default=0.1, description="Distance bin resolution.")
-    unit: LengthUnit = Field(default=LengthUnit.METER, description="Unit.")
-
-    def get(self) -> sc.Variable:
-        return sc.scalar(self.value, unit=self.unit.value)
-
-
-class TimeResolution(BaseModel):
-    """Resolution of the time-of-arrival axis in a TOF lookup table."""
-
-    value: float = Field(default=250.0, description="Time bin resolution.")
-    unit: TimeUnit = Field(default=TimeUnit.MICROSECOND, description="Unit.")
-
-    def get(self) -> sc.Variable:
-        return sc.scalar(self.value, unit=self.unit.value)
-
-
-class LtotalRange(RangeModel):
-    """Range of total flight paths covered by the lookup table."""
-
-    start: float = Field(default=5.0, description="Shortest L_total.")
-    stop: float = Field(default=30.0, description="Longest L_total.")
-    unit: LengthUnit = Field(default=LengthUnit.METER, description="Unit.")
